Add '-tags nodist' to not include the update cmd
Currently with go, if you depend on the net package, you get built
using cgo which results in a binary with dynamic dependencies.

I want a fully static binary so that I can make tiny docker containers
which only have one binary in them, and this facilitates that.

To use, build with `go build -tags nodist`.

Existing builds are unaffected, except that `update` is now the last
command listed in the help.