Simplify Forego.startProcess (#79)
* Remove unnecessary return

* Simplify method Forego.startProcess

We need to call f.teardown.Fall() only when a process has finished
and restart is disabled.

When f.teardown.Barrier() is closed, it’s unnecessary to call
f.teardown.Fall() because we know the barrier is already closed.

* Don’t wait twice on finished

The select has two cases. In the first case, we wait on finished.
We don’t need to wait on finished a second time since we already know
that finished is closed. This is the reason why we moved the defer statement
at the beginning of the second case.

* This goroutine doesn’t need to block until the process has finished