Fix using concurrency and arg together in start
If you ran "forego start" with both arguments and a concurrency flag it
would often hang:

    forego start -c hello=2 world

Secondly, in concurrency mode we didn't check if the process requested
actually existed, which could also cause forego to hang. We fix this two
bugs by unifying the argument and concurrency flag code paths. In the
process, we also extend argument mode to accept multiple processes to
launch, so we can now type:

    forego start hello world

Giving us a slightly nicer interface. If you use both arguments and
concurrency flags as in the first example, this will work and even allow
you to specify the same process to launch in both lists, with the
concurrency value coming from the concurrency list.